requirements.txt
bandit
semgrep
google-genai
openai
anthropic
python-dotenv